dto reponses : models id switched from long type to integer

+++ src/main/java/com/example/immo/dto/responses/RentalResponseDto.java
@@ -11,11 +11,11 @@ import java.util.Date;
 @Setter
 @Getter
 public class RentalResponseDto {
-    private Long id;
+    private Integer id;
     private String description;
     private String name;
     // private ReturnableUserDto owner;
-    private Long owner_id;
+    private Integer owner_id;
     private String picture;
     private Integer surface;
     private Integer price;
